# perf: Improve performance for large bracketed pastes.