Speedup is about an order of magnitude and change (as expected: 23 times faster for 300KB: 93s -> 4s).

The current algorithm has polynomial time complexity. The new algorithm
runs in linear time and is generally much more efficient because it
operates on packed byte strings instead of linked lists of Char.

Addresses part of https://github.com/jtdaugherty/vty/issues/231, but there is more to do.

Some timing:
* 100KB: 1 second
* 200KB: 2.5 seconds
* 300KB: 4 seconds
* 400KB: 7 seconds
* 500KB: 12 seconds
* 600KB: 16 seconds
* 700KB: 22 seconds

<img width="569" alt="image" src="https://user-images.githubusercontent.com/10647936/157513116-f3c2ab58-0fb4-4a10-a0a1-4db7f40edc0c.png">

As we can see, it's still `O(n^2)` overall, probably because of the
calls to `bracketedPasteFinished`. I'll investigate that next. The
constant factor overall is much lower now:
```
Before: 2.866E-6n^3  - 1.784E-4n^2 + 0.114n - 2.622
After:  -1.389E-8n^3 + 5.53E-5n^2  - 1.604E-3n + 0.273
```