@iphydf 2022-03-09
I tried to follow the style of surrounding code. Let me know if there's anything style-wise to fix.

Regarding `start` and `end`: if you're worried about performance of those being inside the function, don't be, they are lifted by the compiler. If style-wise it's not OK, I'm happy to lift them manually, though.