@jtdaugherty 2022-03-21
This change is now released in 5.35.